+++ src/pyforcesim/simulation/base_components.py
@@ -3,7 +3,7 @@ from collections.abc import Generator, Callable
 
 import salabim
 
-
+from pyforcesim.types import Infinite
 if TYPE_CHECKING:
     from pyforcesim.simulation.environment import SimulationEnvironment
 
@@ -31,7 +31,6 @@ class SimulationComponent(salabim.Component):
     
     def run(self) -> Generator[Any, Any, None]:
         """main logic loop for all resources in the simulation environment"""
-        #logger_infstrct.debug(f"----> Process logic of {self}")
         # pre control logic
         ret = self.pre_process()
         # main control logic
@@ -45,3 +44,38 @@ class SimulationComponent(salabim.Component):
         else:
             ret = self.post_process()
 
+class StorageComponent(SimulationComponent):
+    """thin wrapper for Salabim store components to add them as component
+    to simulation entities"""
+    
+    def __init__(
+        self,
+        env: 'SimulationEnvironment',
+        name: str,
+        capacity: int | Infinite,
+        pre_process: Callable[..., Any],
+        sim_logic: Callable[..., Generator[Any, Any, Any]],
+        post_process: Callable[..., Any],
+    ):
+        super().__init__(
+            env=env,
+            name=name,
+            pre_process=pre_process,
+            sim_logic=sim_logic,
+            post_process=post_process,
+        )
+        storage_name = f'{name}_storage'
+        self._store = salabim.Store(
+            env=env,
+            name=storage_name,
+            capacity=capacity, # type: ignore Salabim wrong type hint
+        )
+        self._store_name = self._store.name()
+    
+    @property
+    def store(self) -> salabim.Store:
+        return self._store
+    
+    @property
+    def store_name(self) -> str:
+        return self._store_name
