--- src/pyforcesim/simulation/types.py
@@ -1,5 +1,0 @@
-from typing import NewType
-
-SystemID = NewType('SystemID', int)
-CustomID = NewType('CustomID', str)
-LoadID = NewType('LoadID', int)
